[CH #156498838] Clean up models.py
