+++ api/__init__.py
@@ -3,3 +3,7 @@ from flask import Flask
 app = Flask('__name__')
 
 app.config['JWT_SECRET_KEY'] = 'super-secret-key'
+
+
+users = []
+books = []
