+++ tests/test_models.py
@@ -1,7 +1,8 @@
 import unittest
 import json
 from run import app
-from api.models import *
+from api.models import Book, User
+from api.apis import getAllUsers, getAllBooks, getBook, getBookId, updateBook, updatePassword, deleteAllBooks, deleteBook, borrowBook
 
 
 class TestBooksModel(unittest.TestCase):
